# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/clients": {
      "get": {
        "operationId": "AdminClientsController_listClients",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Legacy offset paging. Prefer cursor.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque keyset cursor (R-2.4).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "withTotal",
            "required": false,
            "in": "query",
            "description": "Counting is a full scan.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search email and name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "individual",
                "referral",
                "partner"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "active",
                "pending",
                "suspended"
              ],
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                0,
                1
              ],
              "type": "number"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "description": "Exact match on the country tag.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "Tag SLUG, not id (ADM-14).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "createdAt",
                "email",
                "firstName",
                "status",
                "type",
                "verificationLevel",
                "country"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Paginated, filterable, sortable client list",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/clients/{id}": {
      "get": {
        "description": "Each section is gated on its own permission and is ABSENT when the caller lacks it — which is deliberately different from present-and-empty, so a screen can distinguish \"hidden from you\" from \"this client has none\".",
        "operationId": "AdminClientsController_getClientProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "A client's full profile — KYC, documents, trading accounts, referrals (FR-ADM-01)",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/permissions": {
      "get": {
        "operationId": "AdminRbacController_getPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/PermissionModuleDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Permission catalog grouped by module (requires roles.view or users.view)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/client-fields": {
      "get": {
        "operationId": "AdminRbacController_listClientFields",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/ClientFieldGroupDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Maskable client fields, grouped (requires roles.view or users.view)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/roles": {
      "get": {
        "operationId": "AdminRbacController_listRoles",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "name",
                "createdAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List RBAC roles (requires roles.view or users.view)",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminRbacController_createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a custom role (requires roles.manage)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/roles/export": {
      "get": {
        "operationId": "AdminRbacController_exportRoles",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "csv"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A CSV file, named `roles-<YYYY-MM-DD>.csv`.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Export the RBAC role definitions as CSV",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/roles/{id}": {
      "put": {
        "operationId": "AdminRbacController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update a custom role (requires roles.manage)",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminRbacController_deleteRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete a custom role (requires roles.manage)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users": {
      "get": {
        "operationId": "AdminRbacController_listAdmins",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "name",
                "email",
                "role",
                "status",
                "createdAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Opting into paging switches the response to the envelope shape.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A bare array by default. With `page` or `limit`, the paginated envelope `{ items, total, page, limit }` instead.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminProfileDto"
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AdminProfileDto"
                          }
                        },
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "items",
                        "total",
                        "page",
                        "limit"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List admin accounts (requires users.view)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/admin-users/export": {
      "get": {
        "operationId": "AdminRbacController_exportAdmins",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "csv"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A CSV file, named `admin-users-<YYYY-MM-DD>.csv`.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Export the administrator directory as CSV",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users/{id}": {
      "patch": {
        "operationId": "AdminRbacController_updateAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update an admin’s name, role, or permissions (requires users.edit)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users/{id}/status": {
      "patch": {
        "description": "Suspension takes effect on the target’s NEXT request — AdminGuard re-reads status on every call — and blocks login. Refused on your own account and on the master admin.",
        "operationId": "AdminRbacController_setAdminStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Suspend or reactivate an administrator (requires users.suspend)",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "OxShare CRM API",
    "description": "Forex/CFD Introducing-Broker CRM — Phase 1",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oxshare_crm_portal_at",
        "description": "Portal session. Set by POST /v1/auth/login; httpOnly, so not settable here."
      },
      "admin": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oxshare_crm_admin_at",
        "description": "Admin session. Set by POST /v1/admin/auth/login; httpOnly, so not settable here."
      }
    },
    "schemas": {
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Logged out."
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ClientListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientRowDto"
            }
          },
          "maskedFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "total": {
            "type": "number",
            "description": "Only when ?withTotal=true. Counting 219,000 rows is a full scan."
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "maskedFields",
          "nextCursor",
          "page",
          "limit"
        ]
      },
...
      "PermissionModuleDto": {
        "type": "object",
        "properties": {
          "moduleName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionItemDto"
            }
          }
        },
        "required": [
          "moduleName",
          "description",
          "permissions"
        ]
      },
...
    }
  }
}